fix: move validateToken to separate module for Next.js route compatibility
Next.js route files can only export HTTP handlers (GET, POST, etc.).
Move validateToken to @/lib/voice-token and import it from there.
